lib: discard uncommon waypoint types in ICAO route parser.
Previously, the code would sometimes select a localizer over e.g.
an identically-named NDB, because the latter was further away from
the last waypoint that said localizer; giving priority to common
waypoint types leads to a more desirable behavior.

We will still fall black to an uncommon type when
there is no match for any of the prioritized types.