YFMS: fix direct to functionality.
We do need to set the tracked leg's origin to the aircraft's
current position, otherwise the plane will simply try to join a
straight line between the previous waypoint in the plan, and the
requested direct termination (which obviously isn't the correct
flight path for a direct to).

We could do it manually, but X-Plane will happily do it for us
via a dedicated command. Literally, all we have to do is set the
displayed FMS entry to the target waypoint and call the command.